+++ src/main/java/com/unzer/payment/Customer.java
@@ -48,7 +48,7 @@ public class Customer implements PaymentType {
 	private String phone;
 	private String mobile;
 	private Address billingAddress;
-	private Address shippingAddress;
+	private ShippingAddress shippingAddress;
 	private String company;
 
 	private CustomerCompanyData companyData;
@@ -155,11 +155,11 @@ public class Customer implements PaymentType {
 		return null;
 	}
 
-	public Address getShippingAddress() {
+	public ShippingAddress getShippingAddress() {
 		return shippingAddress;
 	}
 
-	public Customer setShippingAddress(Address shippingAddress) {
+	public Customer setShippingAddress(ShippingAddress shippingAddress) {
 		this.shippingAddress = shippingAddress;
 		return this;
 	}