+++ src/main/java/com/unzer/payment/ShippingAddress.java
@@ -1,0 +1,40 @@
+package com.unzer.payment;
+
+import com.google.gson.annotations.SerializedName;
+
+public class ShippingAddress extends Address {
+    private Type shippingType;
+
+    public Type getShippingType() {
+        return shippingType;
+    }
+
+    public ShippingAddress setShippingType(Type shippingType) {
+        this.shippingType = shippingType;
+        return this;
+    }
+
+    public static ShippingAddress of(Address address, Type shippingType) {
+        return (ShippingAddress) new ShippingAddress()
+                .setShippingType(shippingType)
+                .setName(address.getName())
+                .setCountry(address.getCountry())
+                .setStreet(address.getStreet())
+                .setCity(address.getCity())
+                .setZip(address.getZip())
+                .setState(address.getState());
+    }
+
+    public enum Type {
+        @SerializedName("equals-billing")
+        EQUALS_BILLING,
+        @SerializedName("different-addresses")
+        DIFFERENT_ADDRESSES,
+        @SerializedName("branch-pickup")
+        BRANCH_PICKUP,
+        @SerializedName("post-office-pickup")
+        POST_OFFICE_PICKUP,
+        @SerializedName("pack-station")
+        PACK_STATION
+    }
+}